refactor(client-lib): implement the certificate client only once
By using private static function so behavior are still implemented in
splitted modules.
